Move funcs to `waiting_parent` module
